functions/requirements.txt
# Azure Functions
azure-functions>=1.17.0
azure-functions-durable>=1.2.0

# Azure Services
azure-search-documents>=11.4.0
azure-storage-blob>=12.19.0
openai>=1.12.0

# Data Processing
pyyaml>=6.0
python-frontmatter>=1.0.0
tiktoken>=0.5.0

# Text Processing
rapidfuzz>=3.5.0  # For citation verification
regex>=2023.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
